Move Plebeian and Vveiss to new UI and delete old UI code
