+++ src/Releases/Release0005_Plebeian/Canvas.js
@@ -1,0 +1,24 @@
+import React, { useEffect, useState } from 'react';
+import useAudioPlayer from '../../UI/Player/hooks/useAudioPlayer';
+import '../Release.css';
+import Scene from './Scene';
+
+export default function Canvas({ hasEnteredWorld }) {
+    const { audioPlayer, currentTrackName } = useAudioPlayer();
+    const [trackName, setTrackName] = useState("Heaven")
+
+    useEffect(() => {
+        if (!currentTrackName) setTrackName("Heaven")
+        else setTrackName(currentTrackName);
+    }, [currentTrackName])
+
+    return <Scene
+        hasEnteredWorld={hasEnteredWorld}
+        trackName={trackName}
+        audioPlayer={audioPlayer}
+    />
+}
+
+
+
+
