+++ src/Releases/Release0006_Vveiss/Canvas.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import useAudioPlayer from '../../UI/Player/hooks/useAudioPlayer';
+import '../Release.css';
+import Scene from './Scene';
+
+export default function Canvas({ }) {
+    const { audioPlayer } = useAudioPlayer();
+    return <Scene audioPlayer={audioPlayer} />
+}
