--- src/UI/LegacyMenu/MenuConstants.js
@@ -1,15 +1,0 @@
-export const SHAPES = [
-  {
-    pathAlt: 'M138.480469,13.1210938 C272.519531,-13.9296875 916.964844,13.1210938 964.617188,22.546875 C971.556974,23.9195863 978.238274,25.1470851 984.661087,26.2293716 L984.661087,26.2293718 C1023.63726,32.797114 1055.08779,61.7061916 1064.90867,99.9923134 C1070.77091,122.845864 1076.37166,148.486447 1081.71094,176.914063 C1093.18341,237.996322 1101.08445,299.358301 1105.41406,361 L1122.78305,572.284544 C1125.92491,610.503672 1106.93395,647.153411 1073.89771,666.625966 C1040.04814,686.577927 998.314009,701.519011 948.695313,711.449219 C869.001917,727.398286 600.169855,716.60543 142.199124,679.070648 L142.199124,679.070648 C139.723053,678.867712 137.255404,678.572574 134.801314,678.185849 L94.1464844,671.779297 C67.57402,667.591901 46.6746488,646.84918 42.2874515,620.308972 C4.79398872,393.493399 -7.16641624,236.899524 6.40625, 150.527344 Z',
-    path: 'M85.828125,73.6445313 C153.05982,41.8952549 329.324219,24.28125 384.433594,61.5273437 C439.542969,98.7734375 915.253312,-22.65625 1012.72266,5.3828125 C1110.192,33.421875 1110.192,53.4492187 1117.65625,84.671875 C1125.1205,115.894531 996.351562,372.002031 1070.27734,476.078125 C1144.20312,580.154219 1053.59375,704.492188 1020.84375,730.583319 C988.09375,756.67445 748.679687,723.179688 567.5,686.28125 C452.776784,662.917111 304.93341,675.06626 123.969877,722.728698 L123.969878,722.7287 C97.2739349,729.759916 68.8265171,724.004834 46.9629181,707.149734 C33.0003893,696.385732 22.2770729,683.002488 14.7929687,667 C-7,620.402344 0.0818498368,590.162379 14.7929687,497.097656 C29.5040877,404.032933 58.046875,400.484375 58.046875,311.835938 C58.046875,223.1875 18.5964302,105.393808 85.828125, 73.6445313 Z',
-    duration: 2000,
-    easing: 'easeInOutSine'
-  },
-
-];
-
-export const MENU_ICON_CLOSE = "M50,100 C22.3857625,100 0,77.6142375 0,50 C0,22.3857625 22.3857625,0 50,0 C77.6142375,0 100,22.3857625 100,50 C100,77.6142375 77.6142375,100 50,100 Z M60.2412109,74.703125 C61.6084053,76.6237075 62.8209583,78.0234331 63.8789062,78.9023438 C64.9368542,79.7812544 66.2307866,80.2207031 67.7607422,80.2207031 C69.4860112,80.2207031 70.9345645,79.7161509 72.1064453,78.7070312 C73.2783262,77.6979116 73.8642578,76.4934966 73.8642578,75.09375 C73.8642578,73.7265557 72.9202568,71.6920708 71.0322266,68.9902344 L58.9716797,51.7050781 L69.6650391,37.3007812 C70.6416064,36.0312437 71.3658831,34.9326218 71.8378906,34.0048828 C72.3098982,33.0771438 72.5458984,32.108729 72.5458984,31.0996094 C72.5458984,29.6998628 71.9681047,28.5035857 70.8125,27.5107422 C69.6568953,26.5178987 68.2002041,26.0214844 66.4423828,26.0214844 C64.9124272,26.0214844 63.6347708,26.4446572 62.609375,27.2910156 C61.5839792,28.137374 60.3714263,29.5370996 58.9716797,31.4902344 L50.5732422,43.3066406 L41.9306641,31.4902344 C40.5634697,29.6022041 39.3671926,28.2187544 38.3417969,27.3398438 C37.3164011,26.4609331 36.0224688,26.0214844 34.4599609,26.0214844 C32.6695874,26.0214844 31.2210342,26.5260366 30.1142578,27.5351562 C29.0074814,28.5442759 28.4541016,29.732415 28.4541016,31.0996094 C28.4541016,32.8574307 29.3818267,34.9244673 31.2373047,37.3007812 L41.9306641,51.7050781 L29.8701172,68.9902344 C28.0471914,71.5944141 27.1357422,73.5963472 27.1357422,74.9960938 C27.1357422,76.3632881 27.7298118,77.5758411 28.9179688,78.6337891 C30.1061257,79.6917371 31.546541,80.2207031 33.2392578,80.2207031 C34.7692134,80.2207031 36.030594,79.7975303 37.0234375,78.9511719 C38.016281,78.1048135 39.228834,76.688812 40.6611328,74.703125 L50.5732422,60.4941406 L60.2412109,74.703125 Z";
-
-export const MENU_ICON_OPEN = "M49.539,0C22.188-0.01,0.018,22.381,0,50.023c0.018,27.601,22.188,49.993,49.539,49.979  c27.344,0.014,49.518-22.378,49.541-49.979C99.057,22.381,76.883-0.01,49.539,0z M49.512,21.634c4.372,0,7.922,3.546,7.922,7.922  s-3.55,7.923-7.922,7.923c-4.38,0-7.927-3.547-7.927-7.923S45.132,21.634,49.512,21.634z M61.117,74.787H38.169v-3.922h3.342  c0.48,0,0.873-0.396,0.873-0.881V44.711c0-0.484-0.389-0.881-0.873-0.881h-3.342v-3.795h18.466v29.567  c0,0.696,0.396,1.263,0.881,1.263h3.602V74.787z";
-
-// export const MENU_INFO_ICON = "M50,100 C22.3857625,100 0,77.6142375 0,50 C0,22.3857625 22.3857625,0 50,0 C77.6142375,0 100,22.3857625 100,50 C100,77.6142375 77.6142375,100 50,100 Z M56.6650391,41.5898438 C56.6650391,39.2135298 56.0302798,37.3580796 54.7607422,36.0234375 C53.4912046,34.6887954 51.8798926,34.0214844 49.9267578,34.0214844 C47.973623,34.0214844 46.3867249,34.6887954 45.1660156,36.0234375 C43.9453064,37.3580796 43.3349609,39.3437368 43.3349609,41.9804688 L43.3349609,80.1152344 C43.3349609,82.7194141 43.9453064,84.7213472 45.1660156,86.1210938 C46.3867249,87.5208403 47.973623,88.2207031 49.9267578,88.2207031 C51.8798926,88.2207031 53.4912046,87.5371162 54.7607422,86.1699219 C56.0302798,84.8027275 56.6650391,82.7845186 56.6650391,80.1152344 L56.6650391,41.5898438 Z M50.0732422,27.8203125 C51.8636157,27.8203125 53.4098242,27.2425188 54.7119141,26.0869141 C56.0140039,24.9313093 56.6650391,23.3281353 56.6650391,21.2773438 C56.6650391,19.1614478 55.997728,17.5501357 54.6630859,16.4433594 C53.3284438,15.336583 51.7985112,14.7832031 50.0732422,14.7832031 C48.2828687,14.7832031 46.7122463,15.3935486 45.3613281,16.6142578 C44.0104099,17.834967 43.3349609,19.3893135 43.3349609,21.2773438 C43.3349609,23.3606875 43.994134,24.9719995 45.3125,26.1113281 C46.630866,27.2506567 48.2177642,27.8203125 50.0732422,27.8203125 Z";
