--- src/UI/LegacyNavigation/Navigation.css
@@ -1,43 +1,0 @@
-.arrow {
-  position: absolute;
-  top: 46%;
-  opacity: 0.5;
-  z-index: 10;
-  width: 50px;
-  height: 50px;
-  transition: 500ms all ease;
-  -webkit-transition: 500ms all ease;
-}
-
-.arrow:hover {
-  opacity: 1;
-  cursor: pointer;
-}
-
-.arrow-prev {
-  left: 0;
-  margin-left: -10px;
-}
-
-.arrow-prev:hover {
-  transform: translate(10px);
-}
-
-.arrow-next {
-  right: 0;
-  margin-right: -2px;
-}
-
-.arrow-next:hover {
-  transform: translate(-10px);
-}
-
-
-@media screen and (max-width: 28em) {
-
-  .arrow {
-    width: 25px;
-    height: 25px;
-  }
-
-}
