--- src/UI/Player/LegacyPlayer.css
@@ -1,240 +1,0 @@
-#player-container {
-    margin: 0%;
-    z-index: 10000;
-    /* position: absolute; */
-    display: -ms-flexbox;
-    bottom: 0;
-    display: flex;
-    -ms-flex-align: center;
-    /* align-items: flex-start;  */
-    align-self: flex-start;
-    margin-right: 20px;
-}
-
-#play-button-container {
-    position: relative;
-    overflow: hidden;
-    top: 10px;
-}
-
-#play-button-container text {
-    font-family: 'Roboto Mono', monospace;
-    font-size: 25px;
-    font-weight: bold;
-    letter-spacing: 9px;
-    fill: #ffffff;
-    stroke: #ffffff;
-}
-
-#play-button-container svg {
-    position: relative;
-    left: -6px;
-    top: 0;
-    bottom: 0;
-    width: 125px;
-    height: 125px;
-    -webkit-animation-name: rotate;
-    -moz-animation-name: rotate;
-    -ms-animation-name: rotate;
-    -o-animation-name: rotate;
-    animation-name: rotate;
-    -webkit-animation-duration: 10s;
-    -moz-animation-duration: 10s;
-    -ms-animation-duration: 10s;
-    -o-animation-duration: 10s;
-    animation-duration: 10s;
-    -webkit-animation-iteration-count: infinite;
-    -moz-animation-iteration-count: infinite;
-    -ms-animation-iteration-count: infinite;
-    -o-animation-iteration-count: infinite;
-    animation-iteration-count: infinite;
-    -webkit-animation-timing-function: linear;
-    -moz-animation-timing-function: linear;
-    -ms-animation-timing-function: linear;
-    -o-animation-timing-function: linear;
-    animation-timing-function: linear;
-}
-
-@-webkit-keyframes rotate {
-    from {
-        -webkit-transform: rotate(360deg);
-    }
-    to {
-        -webkit-transform: rotate(0);
-    }
-}
-
-@-moz-keyframes rotate {
-    from {
-        -moz-transform: rotate(360deg);
-    }
-    to {
-        -moz-transform: rotate(0);
-    }
-}
-
-@-ms-keyframes rotate {
-    from {
-        -ms-transform: rotate(360deg);
-    }
-    to {
-        -ms-transform: rotate(0);
-    }
-}
-
-@-o-keyframes rotate {
-    from {
-        -o-transform: rotate(360deg);
-    }
-    to {
-        -o-transform: rotate(0);
-    }
-}
-
-@keyframes rotate {
-    from {
-        transform: rotate(360deg);
-    }
-    to {
-        transform: rotate(0);
-    }
-}
-
-.button {
-    border: 0;
-    background: transparent;
-    box-sizing: border-box;
-    width: 0;
-    height: 24px;
-    opacity: 0.5;
-    border-color: transparent transparent transparent #red;
-    transition: 500ms all ease;
-    -webkit-transition: 500ms all ease;
-    cursor: pointer;
-    border-style: solid;
-    border-width: 10px 0 12px 20px;
-    position: absolute;
-    margin-top: 50px;
-    left: 48px;
-    outline: none;
-    top: 0;
-}
-
-.button.paused {
-    border-style: double;
-    border-width: 0px 0 0px 20px;
-}
-
-.no-touch .button:hover {
-    opacity: 1;
-    border-color: transparent transparent transparent #FFFFFF;
-}
-
-#playlist-container {
-    z-index: 999999999;
-    position: relative;
-    /* top: -20x; */
-    margin-left: 10px;
-    display: flex;
-    /* align-self: flex-end; */
-    align-items: center;
-}
-
-#playlist-container ul {
-    padding: 0;
-}
-
-#playlist-container .active-track {
-    display: flex;
-    font-weight: bold;
-}
-
-#current-track-smiley {
-    margin-left: -17px;
-    margin-right: 10px;
-}
-
-#playlist li {
-    list-style-type: none;
-    font-size: 13px;
-    padding-bottom: 5px;
-}
-
-.no-touch #playlist li:hover {
-    opacity: 0.5;
-    cursor: pointer;
-}
-
-@media screen and (max-width: 28em) {
-    #play-button-container {
-        top: -10px;
-    }
-    #play-button-container svg {
-        top: 20px;
-        left: -20px;
-    }
-    .button {
-        margin-top: 70px;
-        left: 34px;
-    }
-    #playlist-container {
-        margin-left: -30px;
-        display: flex;
-        align-self: flex-end;
-    }
-    #playlist {
-        max-height: 100px;
-        overflow-y: auto;
-        word-break: break-word;
-    }
-}
-
-/* ADDED from Player demo.. */
-
-/* body {
-    margin: 0;
-    padding: 0;
-    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
-      "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
-      sans-serif;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-  }
-   */
-
-/* code {
-    font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-      monospace;
-  }
-  
-  @media screen and (min-width: 600px) {
-    .container {
-      padding-top: 50px;
-      width: 600px !important;
-    }
-  }
-   */
-/* 
-.box {
-    display: flex !important;
-    align-items: center;
-}
-
-.song-title {
-    margin-left: 20px;
-}
-
-.controls {
-    display: flex !important;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-}
-
-.controls .button {
-    margin: 5px;
-}
-
-.current-track {
-    width: 80%;
-} */
